requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.1
django-appconf==1.0.6
django-cors-headers==4.3.1
django-staticfiles==1.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==23.2
pillow==10.2.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.17.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
